requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
coverage==7.8.2
defusedxml==0.7.1
Django==5.2.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
httpie==3.2.4
idna==3.10
iniconfig==2.1.0
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.4.4
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.1
PyJWT==2.9.0
PySocks==1.7.1
pytest==8.3.5
pytest-cov==6.1.1
pytest-django==4.11.1
python-decouple==3.8
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
setuptools==80.8.0
sqlparse==0.5.3
urllib3==2.4.0
